requirements.txt
asgiref==3.5.0
backports.zoneinfo==0.2.1
cffi==1.15.0
cryptography==36.0.2
Django==4.0.3
django-extensions==3.1.5
django-shell-plus==1.1.7
djangorestframework==3.13.1
psycopg2==2.9.3
pycparser==2.21
PyJWT==2.3.0
pytz==2021.3
sqlparse==0.4.2
